Added connect page
